feat : Add interest certificate APIs
* 흥미 자격증 API 구현
* 목표 달성도 쿼리 결과가 null일경우 예외 처리 구현

+++ src/main/java/com/cos/cercat/learning/repository/StudyTimeLogRepository.java
@@ -9,13 +9,13 @@ import org.springframework.data.repository.query.Param;
 public interface StudyTimeLogRepository extends JpaRepository<StudyTimeLog, Long> {
 
     @Query("SELECT SUM(stl.studyTime) FROM StudyTimeLog stl WHERE stl.goal.id = :goalId")
-    long getStudyTimeSum(@Param("goalId") Long goalId);
+    Long getStudyTimeSum(@Param("goalId") Long goalId);
 
     @Query("""
             SELECT SUM(stl.studyTime) FROM StudyTimeLog stl
             WHERE FUNCTION('DATE_FORMAT', stl.createdAt, '%Y-%m-%d') = CURRENT_DATE()
             AND stl.goal.id = :goalId
             """)
-    int getTodayStudyTime(@Param("goalId") Long goalId);
+    Long getTodayStudyTime(@Param("goalId") Long goalId);
 
 }
